# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SHOPIN",
        "description": "SHOPIN API's integration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List paginated posts",
                "description": "Returns a paginated list of posts including photos as full URLs",
                "operationId": "0f662500b195dc03ad40457524366831",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 30
                                                            },
                                                            "titre": {
                                                                "type": "string",
                                                                "example": "Manteau old style"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Manteau camel Mango"
                                                            },
                                                            "photos": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "http://127.0.0.1:8000/storage/uploads/posts/TqhzM91f7GYgu5V6NwiYYNi1hKZEk9TlA26dZJ8a.webp"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post details",
                "description": "Returns detailed information of a single post including photos and category info",
                "operationId": "efda728ce749fc8efc310bc84af210ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "titre": {
                                                    "type": "string",
                                                    "example": "Manteau old style"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Manteau camel Mango"
                                                },
                                                "photos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "http://127.0.0.1:8000/storage/uploads/posts/TqhzM91f7GYgu5V6NwiYYNi1hKZEk9TlA26dZJ8a.webp"
                                                    }
                                                },
                                                "proprietes": {
                                                    "type": "object",
                                                    "example": {
                                                        "Marque": "Bershka",
                                                        "Taille": "L/42/10",
                                                        "Couleur": "#C19A6B",
                                                        "Article pour": "Femme",
                                                        "Eventuels défauts": "No"
                                                    }
                                                },
                                                "etat": {
                                                    "type": "string",
                                                    "example": "Bon état"
                                                },
                                                "prix": {
                                                    "type": "string",
                                                    "example": "300.00"
                                                },
                                                "sous_categorie_info": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "titre": {
                                                            "type": "string",
                                                            "example": "Manteau"
                                                        },
                                                        "categorie": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "titre": {
                                                                    "type": "string",
                                                                    "example": "Fashion & Vêtements"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Fashion & Vêtements looollmm"
                                                                },
                                                                "icon": {
                                                                    "type": "string",
                                                                    "example": "http://127.0.0.1:8000/storage/uploads/categories/goCedMgqRvvQQ7xzuwNBgraFZSuuvWShzRCknYvP.png"
                                                                },
                                                                "small_icon": {
                                                                    "type": "string",
                                                                    "example": "http://127.0.0.1:8000/storage/uploads/categories/oE8n20szUVwCTh4pSCy8VXSP0KvTliyMBgX2kDxg.png"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        }
    }
}